Closing JESS now close child processes like Jekyll
Working on issue #9. Fix working great on OSX, waiting to be tested on Linux.

+++ src/modules/Jekyll.js
@@ -77,7 +77,9 @@ Jekyll.prototype.serve = function(directory) {
 }
 
 Jekyll.prototype.stop = function() {
-    this.process.kill('SIGINT');
+    if(this.process) {
+        this.process.kill('SIGINT');
+    }
 }
 
 /**
